[MIG] stock_available_sale: Migration to 12.0
The old field requested_date (sale.order) was defined in the
old module sale_order_dates, but in newer versions this was
integrated into the module sale (and the field requested_date
was renamed to commitment_date). So a dependency to sale is
introduced. The problem is: this module sale defines a default
value for the field type (product.attribute) which is set as
mandatory. A test in the module stock_secondary_unit had to be
updated because it created a product attribute without
indicating a value for its type and made the test crash when
run in an integrated way.